requirements.txt
tensorflow
numpy
pandas
scikit-learn
Pillow>=9.3.0
streamlit>=1.16.0
matplotlib>=3.5.0
seaborn>=0.11.2
opencv-python>=4.5.5
